Add unversioned-project my-university
The legacy project depended on an older version of JDesk and many things are still broken and have to be tested at this commit but the general mechanism of this module is already working.

It passes:
- Java 8.
- DB Connection.
- GUI.

It fails on some esoteric details for the business logic I am soon to fix but have to move forward for now.